moved static meshes code to separate file
